third_party/blink/web_tests/external/wpt/credential-management/passwordcredential-framed-get.sub.https-expected.txt

This is a testharness.js-based test.
[FAIL] Same-origin IFrame in same-origin IFrame does not throw.
  assert_equals: expected (object) null but got (string) "NotSupportedError"
Harness: the test ran to completion.
